+++ sdk/src/orderSubscriber/OrderSubscriber.ts
@@ -232,6 +232,7 @@ export class OrderSubscriber {
 	}
 
 	public async unsubscribe(): Promise<void> {
+		this.usersAccounts.clear();
 		await this.subscription.unsubscribe();
 	}
 }