+++ sdk/src/priorityFee/driftPriorityFeeMethod.ts
@@ -1,15 +1,42 @@
 import fetch from 'node-fetch';
-import { HeliusPriorityFeeLevels } from './heliusPriorityFeeMethod';
+import { HeliusPriorityLevel } from './heliusPriorityFeeMethod';
 
-export type DriftPriorityFeeResponse = HeliusPriorityFeeLevels[];
+export type DriftMarketInfo = {
+	marketType: string;
+	marketIndex: number;
+};
+
+export type DriftPriorityFeeLevels = {
+	[key in HeliusPriorityLevel]: number;
+} & {
+	marketType: 'perp' | 'spot';
+	marketIndex: number;
+};
+
+export type DriftPriorityFeeResponse = DriftPriorityFeeLevels[];
 
 export async function fetchDriftPriorityFee(
 	url: string,
-	marketType: string,
-	marketIndex: number
+	marketTypes: string[],
+	marketIndexes: number[]
 ): Promise<DriftPriorityFeeResponse> {
-	const response = await fetch(
-		`${url}?marketType=${marketType}&marketIndex=${marketIndex}`
-	);
-	return await response.json();
+	try {
+		const response = await fetch(
+			`${url}/batchPriorityFees?marketType=${marketTypes.join(
+				','
+			)}&marketIndex=${marketIndexes.join(',')}`
+		);
+		if (!response.ok) {
+			throw new Error(`HTTP error! status: ${response.status}`);
+		}
+		return await response.json();
+	} catch (err) {
+		if (err instanceof Error) {
+			console.error('Error fetching priority fees:', err.message);
+		} else {
+			console.error('Unknown error fetching priority fees:', err);
+		}
+	}
+
+	return [];
 }