+++ sdk/tests/dlob/helpers.ts
@@ -142,6 +142,10 @@ export const mockAMM: AMM = {
 	bidQuoteAssetReserve: new BN(0),
 	askBaseAssetReserve: new BN(0),
 	askQuoteAssetReserve: new BN(0),
+
+	netUnsettledFundingPnl: new BN(0),
+	quoteAssetAmountWithUnsettledLp: new BN(0),
+	referencePriceOffset: 0,
 };
 
 export const mockPerpMarkets: Array<PerpMarketAccount> = [